+++ config/test.exs
@@ -9,8 +9,8 @@ config :brando, Brando.Integration.Endpoint,
 
 config :brando, Brando.Integration.TestRepo,
   url: "ecto://postgres:postgres@localhost/brando_test",
+  types: Brando.PostgresTypes,
   adapter: Ecto.Adapters.Postgres,
-  extensions: [{Postgrex.Extensions.JSON, library: Poison}],
   pool: Ecto.Adapters.SQL.Sandbox,
   ownership_pool: DBConnection.Poolboy,
   pool_overflow: 0