+++ lib/brando/postgrex_types.ex
@@ -1,0 +1,2 @@
+Postgrex.Types.define(Brando.PostgresTypes, [Postgrex.Extensions.JSON] ++
+                                             Ecto.Adapters.Postgres.extensions(), json: Poison)